// src/app/feature/model/substitute.ts
export class Substitute {
    enrollmentId: number;
    oldStudentId: number;
    newStudentId: number;
    maxOut: Date;
    reason: string;
    sponsorId: number;
    firstName: string;
    lastName: string;
    middleInitial: string;
    sponsorAddress: string;
    parishName: string;
    parishCity: string;
}